+++ backend-services/vote-commitment-app/src/main/java/org/cardano/foundation/voting/domain/web3/JwtLoginEnvelope.java
@@ -1,0 +1,22 @@
+package org.cardano.foundation.voting.domain.web3;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class JwtLoginEnvelope {
+
+    private String event;
+
+    private String address;
+
+    private String network;
+
+    private String role;
+
+}