+++ backend-services/voting-app/src/main/java/org/cardano/foundation/voting/service/vote/VoteService.java
@@ -4,7 +4,6 @@ import io.vavr.control.Either;
 import org.cardano.foundation.voting.domain.UserVotes;
 import org.cardano.foundation.voting.domain.VoteReceipt;
 import org.cardano.foundation.voting.domain.entity.Vote;
-import org.cardano.foundation.voting.repository.VoteRepository;
 import org.cardano.foundation.voting.service.auth.jwt.JwtAuthenticationToken;
 import org.cardano.foundation.voting.service.auth.web3.Web3AuthenticationToken;
 import org.zalando.problem.Problem;
@@ -12,8 +11,6 @@ import org.zalando.problem.Problem;
 import java.util.List;
 
 public interface VoteService {
-
-    List<VoteRepository.CompactVote> findAllCompactVotesByEventId(String eventId);
 
     Either<Problem, List<UserVotes>> getVotes(JwtAuthenticationToken auth);