+++ ui/summit-2023/src/pages/Home/Home.tsx
@@ -1,25 +1,28 @@
 import React from 'react';
-import { Typography, Grid } from '@mui/material';
+import { Typography, Grid, useTheme, useMediaQuery } from '@mui/material';
 import CARDANOSUMMIT2023LOGO from '../../common/resources/images/cardanosummit2023.svg';
 import { Hexagon } from '../../components/common/Hexagon';
 import './Home.scss';
 import { i18n } from '../../i18n';
 import { NavLink } from 'react-router-dom';
 import { CustomButton } from '../../components/common/Button/CustomButton';
-import { hasEventEnded } from '../../utils/utils';
 import { useSelector } from 'react-redux';
 import { RootState } from '../../store';
+import { formatUTCDate } from 'utils/dateUtils';
+import Chip from '@mui/material/Chip';
+import EventIcon from '@mui/icons-material/Event';
 
 const Home: React.FC = () => {
   const eventCache = useSelector((state: RootState) => state.user.event);
-  const eventHasEnded = hasEventEnded(eventCache?.eventEndDate);
+  const theme = useTheme();
+  const isMobile = useMediaQuery(theme.breakpoints.down('sm'));
 
   return (
     <Grid
       container
       spacing={1}
       sx={{
-        height: { xs: '60%', md: '70%' },
+        height: { xs: '60%', md: '75vh', lg: '57vh', xl: '71vh' },
         margin: { xs: '0%', sm: '2%', md: '3%', lg: '4%' },
       }}
     >
@@ -32,6 +35,8 @@ const Home: React.FC = () => {
           padding: '20px',
           order: '1',
           display: 'flex',
+          justifyContent: 'center',
+          alignItems: 'center',
         }}
       >
         <div className="left-title-container">
@@ -42,12 +47,34 @@ const Home: React.FC = () => {
           >
             {i18n.t('landing.title')}
           </Typography>
+          {isMobile ? (
+            <div className="event-time">
+                <p>Opens on: {formatUTCDate(eventCache?.eventStartDate?.toString())}</p>
+                <p>Closes on: {formatUTCDate(eventCache?.eventEndDate?.toString())}</p>
+            </div>
+          ) : (
+            <Chip
+              sx={{
+                height: '46px',
+                borderRadius: '8px',
+                my: '20px',
+                px: '10px',
+              }}
+              icon={<EventIcon />}
+              label={`The Vote opens on ${formatUTCDate(
+                eventCache?.eventStartDate?.toString()
+              )}, and closes on ${formatUTCDate(eventCache?.eventEndDate?.toString())}.`}
+              color="primary"
+            />
+          )}
+
           <Typography
             variant="body1"
             sx={{ textAlign: { xs: 'center', sm: 'left' } }}
           >
             {i18n.t('landing.description')}
           </Typography>
+
           <Grid
             container
             sx={{ justifyContent: { xs: 'center', sm: 'left' } }}
@@ -60,10 +87,10 @@ const Home: React.FC = () => {
                 styles={{
                   background: '#ACFCC5',
                   color: '#03021F',
-                  marginTop: '40px',
+                  marginTop: '20px',
                   textDecoration: 'none !important',
                 }}
-                label={eventHasEnded ? 'Voting ended' : i18n.t('landing.getStartedButton')}
+                label={eventCache?.finished ? 'Voting ended' : i18n.t('landing.getStartedButton')}
               />
             </NavLink>
           </Grid>